chore: merge validation service unit tests
